Refactor: use forward declaration and ReflectorClientSpp.cpp instead of LATRY_SERVICE_BUILD guards
